Added in AlphaNumeric Encryption "AlphaNum"
I added in the AlphaNumeric encryption which is inthe Encrypt.java class. Hopefully I will be able to get an Java GUI for at least the ceaser cypher released soon. This update is for the first initial release. I know I jumped through a lot of numbers but considering where this program is at now, it feels like an full release for an console program.